更新了合约

+++ contracts/ZzbusdtPriceOracleInterface.sol
@@ -1,5 +1,5 @@
 pragma solidity >=0.4.22 <0.9.0;
 
 interface ZzbusdtPriceOracleInterface {
-  function getLatestZzbusdtPrice() external returns (uint);
+  function getLatestZzbusdtPrice() external returns (uint256);
 }
